Accept non-base64 data URIs, keep SVG intrinsic sizes exact, flag inline SVG
Three follow-ups on SVG support, each found by checking rather than assuming.

**data: URIs.** Only `;base64` payloads were accepted; anything else was
rejected outright. That happens to be the *usual* way to inline an SVG
(`data:image/svg+xml,%3Csvg...%3E`), so the common form failed while the rare
one worked. Non-base64 payloads are now percent-decoded per RFC 2397, in
`<img src>` and CSS `url()` alike. Tabs and newlines are dropped from the
payload (as the URL standard does) but spaces are kept — they separate tokens
in an unencoded SVG. Raster images are unaffected in practice; they were
already always base64.

**object-fit.** It already worked for SVG, since `object_fit_rect` only needs an
intrinsic size — but that size was being rounded to integers. An SVG's is
allowed to be fractional, and 40.6x10.4 rounded to 41x10 shifts the aspect
ratio by 5%: `contain` produced a height of 24.4 instead of 25.6, and the whole
drawing was skewed to fit a wrongly-shaped rect. `PreparedImage`'s intrinsic
size is now f32 end-to-end. This also simplifies `BackgroundImagePaint`, which
stored u32 only to cast back to f32 at its one use. Raster sizes are unchanged —
whole pixels either way — and `/Width`/`/Height` round back at the one place
that needs integers.

**Inline SVG stays out of scope, but says so.** `svg { display: none }` in the
UA stylesheet drops the subtree, which was fine when nothing SVG rendered.
Now that `<img src="*.svg">` works, silence is misleading. One warning per
document (per document, not per process — the gem and server render many),
counting what it found. Supporting inline SVG means rebuilding SVG XML from the
HTML DOM and deciding how attribute case, CSS inheritance and `currentColor`
carry across; that is a different problem from referencing a file.

Tests: +6 object-fit/object-position (including the fractional-ratio
regression), +3 inline SVG, +5 data URI forms, +6 unit tests for the decoder.
The rounding and warning fixes were each confirmed load-bearing by reverting
them and watching the right tests fail.

Verified: 1282 tests pass on default features, 1288 with svg-text, clippy clean
in three feature configs, gem specs 141/141, and all five object-fit modes
checked visually against a fractional-size SVG.

+++ core/src/layout/table.rs
@@ -783,7 +783,7 @@ fn compute_natural_content_width(
         BoxContent::Image(image_content) => image_content
             .attr_width
             .map(|w| w as f32)
-            .or_else(|| image_content.image.as_ref().map(|img| img.width as f32))
+            .or_else(|| image_content.image.as_ref().map(|img| img.width))
             .unwrap_or(0.0),
     }
 }
